+++ android/src/main/java/com/smileidentity/react/viewmanagers/SmileIDEnhancedDocumentVerificationViewManager.kt
@@ -47,6 +47,7 @@ class SmileIDEnhancedDocumentVerificationViewManager(
       view.idAspectRatio = it.getFloatOrDefault("idAspectRatio")
       view.allowNewEnroll = it.getBoolOrDefault("allowNewEnroll", false)
       view.skipApiSubmission = it.getBoolOrDefault("skipApiSubmission", false)
+      view.useStrictMode = it.getBoolOrDefault("useStrictMode", false)
     }
   }