+++ ios/Utils/SmileIDUtils.swift
@@ -8,3 +8,12 @@ extension String {
         return false
     }
 }
+
+func getCurrentIsoTimestamp() -> String {
+    let pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
+    let formatter = DateFormatter()
+    formatter.dateFormat = pattern
+    formatter.locale = Locale(identifier: "en_US")
+    formatter.timeZone = TimeZone(identifier: "UTC")
+    return formatter.string(from: Date())
+}