+++ ios/View/SmileIDSmartSelfieCaptureView.swift
@@ -66,6 +66,6 @@ extension SmileIDSmartSelfieCaptureView: SmartSelfieResultDelegate {
     }
 
     func didError(error: Error) {
-        smileIDUIViewDelegate.onError(error: error.localizedDescription)
+        smileIDUIViewDelegate.onError(error: error)
     }
 }