+++ ios/View/SmileIDSmartSelfieEnrollmentEnhancedView.swift
@@ -40,6 +40,6 @@ extension SmileIDSmartSelfieEnrollmentEnhancedView: SmartSelfieResultDelegate {
     }
 
     func didError(error: Error) {
-        smileIDUIViewDelegate.onError(error: error.localizedDescription)
+        smileIDUIViewDelegate.onError(error: error)
     }
 }