+++ ios/View/SmileIDSmartSelfieEnrollmentView.swift
@@ -42,6 +42,6 @@ extension SmileIDSmartSelfieEnrollmentView: SmartSelfieResultDelegate {
     }
 
     func didError(error: Error) {
-        smileIDUIViewDelegate.onError(error: error.localizedDescription)
+        smileIDUIViewDelegate.onError(error: error)
     }
 }