+++ src/SmileIDBiometricKYCView.tsx
@@ -4,22 +4,15 @@ import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNati
 import type { BiometricKYCRequest } from './index';
 import { useSmileIDView } from './useSmileIDView';
 
-const defaultConsentInfo = {
-  consentInformation: {
-    timestamp: new Date().toISOString(),
-  },
-};
-
 const SmileIDBiometricKYCComponent =
   codegenNativeComponent<BiometricKYCRequest>(
     'SmileIDBiometricKYCView'
   ) as HostComponent<BiometricKYCRequest>;
 
 const SmileIDBiometricKYCView: React.FC<BiometricKYCRequest> = (props) => {
-  const mergedProps = { ...defaultConsentInfo, ...props };
-  const viewRef = useSmileIDView('SmileIDBiometricKYCView', mergedProps);
+  const viewRef = useSmileIDView('SmileIDBiometricKYCView', props);
 
-  return <SmileIDBiometricKYCComponent ref={viewRef} {...mergedProps} />;
+  return <SmileIDBiometricKYCComponent ref={viewRef} {...props} />;
 };
 
 export default SmileIDBiometricKYCView;