+++ src/types.ts
@@ -129,7 +129,21 @@ export type DocumentVerificationRequest = SmartSelfieRequest & {
   /*jobId The job ID to associate with the job. Most often, this will correspond to a unique
   Job ID within your own system. If not provided, a random job ID will be generated.*/
   jobId?: string;
+
+  /*
+   * Will use Enhanced SmartSelfie™ capture for the selfie capture
+   * if set to true, if false will use the default SmartSelfie™ capture
+   */
+  useStrictMode?: boolean;
 };
+
+export type EnhancedDocumentVerificationRequest =
+  DocumentVerificationRequest & {
+    /**
+     * The country of issuance of the ID type to be captured.
+     */
+    consentInformation: ConsentInformation;
+  };
 
 export type ConsentRequest = Omit<SmartSelfieRequest, 'allowAgentMode'> & {
   /**
@@ -151,10 +165,20 @@ export type ConsentRequest = Omit<SmartSelfieRequest, 'allowAgentMode'> & {
   productName: string;
 };
 
-export type BiometricKYCRequest = ConsentRequest & {
+export type BiometricKYCRequest = SmartSelfieRequest & {
+  /*
+   * The user id information for KYC
+   */
   idInfo: IdInfo;
-  productName: string;
-  jobId: string;
+  /*
+   * The consent information for the partner
+   */
+  consentInformation: ConsentInformation;
+  /*
+   * Will use Enhanced SmartSelfie™ capture for the selfie capture
+   * if set to true, if false will use the default SmartSelfie™ capture
+   */
+  useStrictMode?: boolean;
 };
 
 // noinspection JSUnusedGlobalSymbols
@@ -311,6 +335,41 @@ export class PrepUploadResponse {
     this.refId = refId;
     this.uploadUrl = uploadUrl;
     this.smileJobId = smileJobId;
+  }
+}
+
+/*
+ * The consent information for the user
+ * required for Biometric KYC and Enhanced Document Verification
+ */
+export class ConsentInformation {
+  /*
+   * The date the user granted consent
+   */
+  consentGrantedDate: string;
+  /*
+   * Whether the user has granted consent for personal details
+   */
+  personalDetailsConsentGranted: boolean;
+  /*
+   * Whether the user has granted consent for contact information
+   */
+  contactInfoConsentGranted: boolean;
+  /*
+   * Whether the user has granted consent for document information
+   */
+  documentInfoConsentGranted: boolean;
+
+  constructor(
+    consentGrantedDate: string,
+    personalDetailsConsentGranted: boolean,
+    contactInfoConsentGranted: boolean,
+    documentInfoConsentGranted: boolean
+  ) {
+    this.consentGrantedDate = consentGrantedDate;
+    this.personalDetailsConsentGranted = personalDetailsConsentGranted;
+    this.contactInfoConsentGranted = contactInfoConsentGranted;
+    this.documentInfoConsentGranted = documentInfoConsentGranted;
   }
 }